UpdateDialog string format param
Added optional string format param (regex) to UpdateDialog constructor.
This will allow me use UpdateDialog in the future for data entry for
other things without being constrained to the version string format that
was previously the default.

Also, I realized that I mixed up some commits earlier with LabelChange
and the cancel message for package creation but I moved LabelChange
functionality and removed the method anyway so I guess oops for that but
I fixed it?